refactor: Update import path for useHandleRoute hook and networkHelpers functions
