+++ apps/launch/common-util/functions/networkHelpers.ts
